httpfs: Allow specifying BaseURL in JSON blob
This commit changes the FileTreeRoot JSON format so that a new key
`$version` is supported. When `$version = 2`, the JSON blob can contain
a `baseURL` key that specifies the base URL for all file path. This
allows the httpfs to be in a different domain than the server without
having to use a reverse proxy.